LayoutTests/imported/w3c/web-platform-tests/svg/animations/animateMotion-calcMode-discrete-keyTimes-expected.txt


FAIL Test animateMotion with calcMode="discrete", path, and non-uniform keyTimes assert_equals: point 1 (x) expected 0 but got 25
